sort failed/timed out domains
